requirements.txt
python-telegram-bot==20.8
quart
uvicorn[standard]